Changes to how times are handled in the web app. Closes #26 (I think).
Shift stringifying of reading created at from server to client. Client
now has a dependency on moment.js.
Previous commit means that all timestamps and keys will be UTC time. Now
when we receive timestamps from the DB and API we're receiving UTC
milliseconds since epoch. This feeds into moment's utc() parsing which
requires milliseconds rather than seconds. In the client we've got a new
angular filter that turns the timestamp into a relative humanised
string. This is local to the user, after parsing the UTC timestamp. I
think this covers it, but these timezone things hurt my head, especially
when thinking about server time, buoy time and user time. Currently
ignoring buoy time, which happens to be user time because I'm the only
user and the buoys are all in my timezone :)